+++ tests/src/test_amdgpuinfopmoverdrive.cpp
@@ -108,6 +108,22 @@ TEST_CASE("GPUInfoPMOverdrive tests",
     REQUIRE(output.front() == ::AMD::GPUInfoPMOverdrive::Clk);
   }
 
+  SECTION("Provides clock offset capability")
+  {
+    // clang-format off
+    std::vector<std::string> ppOdClkVoltageData{"OD_SCLK_OFFSET:",
+                                                "-300MHz"};
+    // clang-format on
+
+    ::AMD::GPUInfoPMOverdrive ts(std::make_unique<VectorStringPathDataSourceStub>(
+        "pp_od_clk_voltage", std::move(ppOdClkVoltageData)));
+
+    auto output = ts.provideCapabilities(vendor, gpuIndex, path);
+
+    REQUIRE(output.size() == 1);
+    REQUIRE(output.front() == ::AMD::GPUInfoPMOverdrive::ClkOffset);
+  }
+
   SECTION("Provides voltage offset capability")
   {
     // clang-format off