+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Allow custom Polkit policy files install directory (#477).
 - Fan stop and fan stop temperature controls for AMD RX 7000 series. Requires Linux 6.13 or later.
+- Frequency offset control for AMD RX 9000 series (#481). Requires Linux 6.14 or later.
 - New command line options:
   - `--activate-manual-profile`: activate the given manual profile (#448).
   - `--deactivate-manual-profile`: deactivate the given manual profile (#448).